+++ test/test_ext_show_clock_synchronization.py
@@ -1,5 +1,5 @@
 from flockwave.server.ext.show.clock import ClockSynchronizationHandler, ShowClock
-from pytest import fixture
+from pytest import fixture, raises
 from time import time
 from typing import Tuple
 
@@ -50,7 +50,7 @@ def test_synchronization_basic(clocks: ClockPair, handler: ClockSynchronizationH
     handler.secondary_clock = secondary
     assert not secondary.running
 
-    handler.synchronize_to(primary, primary.ticks_per_second * 10)
+    handler.synchronize_to(primary, 10)
     assert secondary.running
     assert secondary.start_time == primary.start_time + 10
     assert secondary.ticks_given_time(now + 10) == 0
@@ -64,7 +64,7 @@ def test_sync_stop_start(clocks: ClockPair, handler: ClockSynchronizationHandler
     primary.start_time = now
     handler.secondary_clock = secondary
 
-    handler.synchronize_to(primary, primary.ticks_per_second * 10)
+    handler.synchronize_to(primary, 10)
     assert secondary.running
     assert secondary.ticks_given_time(now + 10) == 0
 
@@ -88,7 +88,7 @@ def test_detach_primary_clock(clocks: ClockPair, handler: ClockSynchronizationHa
     primary.start_time = now
     handler.secondary_clock = secondary
 
-    handler.synchronize_to(primary, primary.ticks_per_second * 10)
+    handler.synchronize_to(primary, 10)
     assert handler.primary_clock is primary
     assert secondary.running
 
@@ -100,3 +100,47 @@ def test_detach_primary_clock(clocks: ClockPair, handler: ClockSynchronizationHa
     primary.start_time = now - 5
     assert not secondary.running
     assert secondary.ticks_given_time(now) == 0
+
+
+def test_detach_primary_clock_without_stopping_secondary_clock(
+    clocks: ClockPair, handler: ClockSynchronizationHandler
+):
+    primary, secondary = clocks
+
+    now = time()
+    primary.start_time = now
+    handler.secondary_clock = secondary
+
+    handler.synchronize_to(primary, 10)
+    assert handler.primary_clock is primary
+    assert secondary.running
+
+    handler.disable()
+    assert handler.primary_clock is None
+    assert secondary.running
+    assert secondary.ticks_given_time(now) == -10 * secondary.ticks_per_second
+
+
+def test_secondary_clock_context_manager(
+    clocks: ClockPair, handler: ClockSynchronizationHandler
+):
+    assert handler.secondary_clock is None
+
+    with handler.use_secondary_clock(clocks[1]):
+        assert handler.secondary_clock is clocks[1]
+
+    assert handler.secondary_clock is None
+
+    try:
+        with handler.use_secondary_clock(clocks[1]):
+            assert handler.secondary_clock is clocks[1]
+            raise RuntimeError("foo")
+    except RuntimeError:
+        pass
+
+    assert handler.secondary_clock is None
+
+    with handler.use_secondary_clock(clocks[1]):
+        with raises(RuntimeError):
+            with handler.use_secondary_clock(clocks[0]):
+                pass